Updated a lot

+++ libft/ft_strsub.c
@@ -6,13 +6,13 @@
 /*   By: vbrazas <vbrazas@student.unit.ua>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/11/05 20:38:54 by vbrazas           #+#    #+#             */
-/*   Updated: 2017/11/12 18:26:17 by vbrazas          ###   ########.fr       */
+/*   Updated: 2018/05/24 20:40:09 by vbrazas          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-char	*ft_strsub(char const *s, unsigned int start, size_t len)
+char	*ft_strsub(char const *s, size_t start, size_t len)
 {
 	char	*buf;
 	size_t	i;
